Introduce task runner
