+++ extensions/kafka-connect/src/main/java/com/hazelcast/jet/kafka/connect/impl/TaskRunner.java
@@ -1,0 +1,160 @@
+/*
+ * Copyright 2023 Hazelcast Inc.
+ *
+ * Licensed under the Hazelcast Community License (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://hazelcast.com/hazelcast-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hazelcast.jet.kafka.connect.impl;
+
+import com.hazelcast.logging.ILogger;
+import com.hazelcast.logging.Logger;
+import org.apache.kafka.connect.source.SourceConnector;
+import org.apache.kafka.connect.source.SourceRecord;
+import org.apache.kafka.connect.source.SourceTask;
+import org.apache.kafka.connect.source.SourceTaskContext;
+import org.apache.kafka.connect.storage.OffsetStorageReader;
+
+import javax.annotation.Nonnull;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+
+import static com.hazelcast.jet.impl.util.ExceptionUtil.rethrow;
+
+class TaskRunner {
+    private static final ILogger LOGGER = Logger.getLogger(TaskRunner.class);
+    private static final int AWAIT_STOP_TIMEOUT = 1_000;
+    private final SourceConnector connector;
+    private final AtomicBoolean running = new AtomicBoolean();
+    private CountDownLatch taskStoppedLatch = new CountDownLatch(1);
+    private SourceTask task;
+    private volatile boolean reconfigurationRequested;
+    private final Map<Map<String, ?>, Map<String, ?>> partitionsToOffset;
+
+    TaskRunner(SourceConnector connector, Map<Map<String, ?>, Map<String, ?>> partitionsToOffset) {
+        this.connector = connector;
+        this.partitionsToOffset = partitionsToOffset;
+    }
+
+    List<SourceRecord> poll() {
+        restartTaskIfNeeded();
+        return doPoll();
+    }
+
+    void stop() {
+        LOGGER.info("Stopping task if running");
+        if (running.compareAndSet(true, false)) {
+            try {
+                if (task != null) {
+                    LOGGER.fine("Stopping task");
+                    task.stop();
+                    LOGGER.fine("Task stopped");
+                }
+            } finally {
+                taskStoppedLatch.countDown();
+                taskStoppedLatch = new CountDownLatch(1);
+                task = null;
+            }
+        }
+    }
+
+    void requestReconfiguration() {
+        LOGGER.info("Task reconfiguration requested");
+        reconfigurationRequested = true;
+    }
+
+    private List<SourceRecord> doPoll() {
+        try {
+            return task.poll();
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw rethrow(e);
+        }
+    }
+
+    private void restartTaskIfNeeded() {
+        if (reconfigurationRequested) {
+            reconfigurationRequested = false;
+            stopTaskAndAwait();
+        }
+        if (running.get()) {
+            LOGGER.info("Not starting task, already running");
+            return;
+        }
+        startTaskIfNotRunning();
+    }
+
+    private void startTaskIfNotRunning() {
+        LOGGER.info("Starting tasks if the previous not running");
+        if (running.compareAndSet(false, true)) {
+            LOGGER.info("Starting tasks");
+
+            List<Map<String, String>> taskConfigs = connector.taskConfigs(1);
+            if (!taskConfigs.isEmpty()) {
+                Map<String, String> taskConfig = taskConfigs.get(0);
+                this.task = newInstance(connector.taskClass().asSubclass(SourceTask.class));
+                LOGGER.info("Initializing task: " + task);
+                task.initialize(new JetSourceTaskContext(taskConfig));
+                LOGGER.info("Starting task: " + task + " with task config: " + taskConfig);
+                task.start(taskConfig);
+            } else {
+                LOGGER.info("No task configs!");
+            }
+        }
+    }
+
+    @Nonnull
+    private static <T> T newInstance(Class<? extends T> clazz) {
+        try {
+            return clazz.getConstructor().newInstance();
+        } catch (Exception e) {
+            throw rethrow(e);
+        }
+    }
+
+    private void stopTaskAndAwait() {
+        stop();
+        if (!awaitStop()) {
+            LOGGER.info("Waiting for stopping task timed-out");
+        }
+    }
+
+    private boolean awaitStop() {
+        try {
+            LOGGER.info("Waiting for stopping the previous task");
+            return taskStoppedLatch.await(AWAIT_STOP_TIMEOUT, TimeUnit.MILLISECONDS);
+        } catch (InterruptedException e) {
+            return false;
+        }
+    }
+
+    private final class JetSourceTaskContext implements SourceTaskContext {
+        private final Map<String, String> taskConfig;
+
+        private JetSourceTaskContext(Map<String, String> taskConfig) {
+            this.taskConfig = taskConfig;
+        }
+
+        @Override
+        public Map<String, String> configs() {
+            return taskConfig;
+        }
+
+        @Override
+        public OffsetStorageReader offsetStorageReader() {
+            return new SourceOffsetStorageReader(partitionsToOffset);
+        }
+    }
+}
